Avatar lurping out of position upon play
Issue with placing an avatar at the position i want - but as soon as i play the animation, it snaps avatar back to 0, 0, 0. How do i get the character to stay where I want it and not snap back? Is there a way to to get rid of what appears to be a baked-into-position from the animation file? (see video example)